Rename to workaround Sandbox
