+++ VoodooInput/Info.plist
@@ -291,7 +291,7 @@
 			<key>MT1Props</key>
 			<dict>
 				<key>IOUserClientClass</key>
-				<string>VoodooInputWellspringUserClient</string>
+				<string>IOHIDVoodooInputWellspringUserClient</string>
 				<key>parser-type</key>
 				<integer>1000</integer>
 				<key>parser-options</key>
